requirements.txt
yara-python==3.11.0
Flask==1.1.1